support chromosome 23 and 24 in mutation mapper
